+++ src/navigation/HowItWorksStack.tsx
@@ -1,22 +1,98 @@
 import React, { FunctionComponent } from "react"
+import { ImageSourcePropType } from "react-native"
+import { useTranslation } from "react-i18next"
+import { useNavigation } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
 
-import { Stack as DestinationStack } from "./index"
+import {
+  HowItWorksScreen as HowItWorksScreenType,
+  HowItWorksScreens,
+  Stack as StackType,
+  Stacks,
+} from "../navigation/index"
 import HowItWorksScreen from "../HowItWorks/HowItWorksScreen"
-import useHowItWorksData, {
-  HowItWorksScreenDatum,
-} from "../HowItWorks/useHowItWorksData"
+
+import { Images } from "../assets"
 
 const Stack = createStackNavigator()
 
 interface HowItWorksStackProps {
-  destinationOnSkip: DestinationStack
+  destinationOnSkip: StackType
+}
+
+type HowItWorksScreenDatum = {
+  name: HowItWorksScreenType
+  image: ImageSourcePropType
+  imageLabel: string
+  header: string
+  primaryButtonLabel: string
+  primaryButtonOnPress: () => void
 }
 
 const HowItWorksStack: FunctionComponent<HowItWorksStackProps> = ({
   destinationOnSkip,
 }) => {
-  const howItWorksData = useHowItWorksData(destinationOnSkip)
+  const { t } = useTranslation()
+  const navigation = useNavigation()
+
+  const howItWorksData = (
+    destinationOnSkip: StackType = Stacks.Home,
+  ): HowItWorksScreenDatum[] => {
+    const introduction: HowItWorksScreenDatum = {
+      name: HowItWorksScreens.Introduction,
+      image: Images.PeopleHighFiving,
+      imageLabel: t("onboarding.screen1_image_label"),
+      header: t("onboarding.screen1_header"),
+      primaryButtonLabel: t("onboarding.screen1_button"),
+      primaryButtonOnPress: () =>
+        navigation.navigate(HowItWorksScreens.PhoneRemembersDevices),
+    }
+    const phoneRemembersDevices: HowItWorksScreenDatum = {
+      name: HowItWorksScreens.PhoneRemembersDevices,
+      image: Images.PeopleOnPhones,
+      imageLabel: t("onboarding.screen2_image_label"),
+      header: t("onboarding.screen2_header"),
+      primaryButtonLabel: t("onboarding.screen2_button"),
+      primaryButtonOnPress: () =>
+        navigation.navigate(HowItWorksScreens.PersonalPrivacy),
+    }
+    const personalPrivacy: HowItWorksScreenDatum = {
+      name: HowItWorksScreens.PersonalPrivacy,
+      image: Images.PersonWithLockedPhone,
+      imageLabel: t("onboarding.screen3_image_label"),
+      header: t("onboarding.screen3_header"),
+      primaryButtonLabel: t("onboarding.screen3_button"),
+      primaryButtonOnPress: () =>
+        navigation.navigate(HowItWorksScreens.GetNotified),
+    }
+    const getNotified: HowItWorksScreenDatum = {
+      name: HowItWorksScreens.GetNotified,
+      image: Images.PersonGettingNotification,
+      imageLabel: t("onboarding.screen4_image_label"),
+      header: t("onboarding.screen4_header"),
+      primaryButtonLabel: t("onboarding.screen4_button"),
+      primaryButtonOnPress: () =>
+        navigation.navigate(HowItWorksScreens.ValueProposition),
+    }
+    const valueProposition: HowItWorksScreenDatum = {
+      name: HowItWorksScreens.ValueProposition,
+      image: Images.PersonAndHealthExpert,
+      imageLabel: t("onboarding.screen5_image_label"),
+      header: t("onboarding.screen5_header"),
+      primaryButtonLabel: t("onboarding.screen_5_button"),
+      primaryButtonOnPress: () => navigation.navigate(destinationOnSkip),
+    }
+
+    const howItWorksScreensScreensData: HowItWorksScreenDatum[] = [
+      introduction,
+      phoneRemembersDevices,
+      personalPrivacy,
+      getNotified,
+      valueProposition,
+    ]
+
+    return howItWorksScreensScreensData
+  }
 
   const toStackScreen = (datum: HowItWorksScreenDatum, idx: number) => {
     const screenNumber = idx + 1
@@ -34,6 +110,7 @@ const HowItWorksStack: FunctionComponent<HowItWorksStackProps> = ({
           <HowItWorksScreen
             {...props}
             howItWorksScreenContent={howItWorksScreenDisplayDatum}
+            totalScreenCount={howItWorksData().length}
             destinationOnSkip={destinationOnSkip}
           />
         )}
@@ -43,7 +120,7 @@ const HowItWorksStack: FunctionComponent<HowItWorksStackProps> = ({
 
   return (
     <Stack.Navigator headerMode="none">
-      {howItWorksData.map((data, idx) => toStackScreen(data, idx))}
+      {howItWorksData().map((data, idx) => toStackScreen(data, idx))}
     </Stack.Navigator>
   )
 }
